Update
added a if statement to track the best score. added a variable of play_count to keep track of the best score instead of resetting each time there was a new game. added guess_count +=1 under the last elif to properly keep track of the amount of attempts there are after each game reset.